requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
flask>=2.0.0
picamera2>=0.3.12
scikit-image>=0.19.0
pillow>=9.0.0
matplotlib>=3.5.0
pyserial>=3.5 